app/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.119.1
<<<<<<< HEAD
googlemaps==4.10.0
h11==0.16.0
httptools==0.7.1
=======
genson==1.3.0
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
>>>>>>> 9d6a451d13211a74de66aeda9c0c76c3f34897e0
idna==3.11
polyline==2.0.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
PyYAML==6.0.3
<<<<<<< HEAD
=======
redis==7.0.1
>>>>>>> 9d6a451d13211a74de66aeda9c0c76c3f34897e0
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
<<<<<<< HEAD
=======
geopandas
geoalchemy2
psycopg2-binary
sqlalchemy
asyncpg
>>>>>>> 9d6a451d13211a74de66aeda9c0c76c3f34897e0
